+++ package.json
@@ -21,6 +21,7 @@
   },
   "devDependencies": {
     "@vue/component-compiler-utils": "^3.1.2",
+    "sass": "^1.26.5",
     "vue-template-compiler": "^2.6.11"
   }
 }
