+++ src/game.js
@@ -1,4 +1,4 @@
 import Vue from 'vue';
-import Game from './Game.vue';
+import Game from './pages/host/Game.vue';
 import 'bulma';
 new Vue({ render: createElement => createElement(Game) }).$mount('#app');
