+++ src/play.js
@@ -1,4 +1,4 @@
 import Vue from 'vue';
-import Play from './Play.vue';
+import Play from './pages/player/Play.vue';
 import 'bulma';
 new Vue({ render: createElement => createElement(Play) }).$mount('#app');
